refactor: rename hlp to common in hlp package
