+++ hyperbench/hlp/__init__.py
@@ -3,7 +3,7 @@ from .gcn_hlp import GCNEncoderConfig, GCNHlpModule
 from .hgnn_hlp import HGNNHlpModule, HGNNEncoderConfig
 from .hnhn_hlp import HNHNEncoderConfig, HNHNHlpModule
 from .hgnnp_hlp import HGNNPEncoderConfig, HGNNPHlpModule
-from .hlp import HlpModule
+from .common import HlpModule
 from .hypergcn_hlp import HyperGCNHlpModule, HyperGCNEncoderConfig
 from .mlp_hlp import MLPHlpModule, MlpEncoderConfig
 from .node2vec_common import (
